requirements.txt
pandas==2.2.2
sqlalchemy==2.0.30
python-dotenv==1.0.1
pyodbc==5.1.0